+++ src/libgeodecomp/storage/test/unit/defaultfiltertest.h
@@ -14,13 +14,13 @@ public:
     void testHostAoS()
     {
         std::vector<TestCell<2> > hostCellVec(40);
-        std::vector<double> hostBuffer(40, -1);
+        std::vector<float> hostBuffer(40, -1);
 
         for (std::size_t i = 0; i < hostCellVec.size(); ++i) {
             hostCellVec[i].testValue = i + 0.5;
         }
 
-        FilterBase<TestCell<2> > *filter = new DefaultFilter<TestCell<2>, double, double>();
+        FilterBase<TestCell<2> > *filter = new DefaultFilter<TestCell<2>, float, float>();
         char TestCell<2>::* memberPointer = reinterpret_cast<char TestCell<2>::*>(&TestCell<2>::testValue);
 
         filter->copyMemberOut(
@@ -48,20 +48,20 @@ public:
             memberPointer);
 
         for (std::size_t i = 0; i < hostCellVec.size(); ++i) {
-            TS_ASSERT_EQUALS(47.11 + i, hostCellVec[i].testValue);
+            TS_ASSERT_EQUALS(float(47.11 + i), hostCellVec[i].testValue);
         }
     }
 
     void testHostSoA()
     {
-        std::vector<double> hostMemberVec(40);
-        std::vector<double> hostBuffer(40, -1);
+        std::vector<float> hostMemberVec(40);
+        std::vector<float> hostBuffer(40, -1);
 
         for (std::size_t i = 0; i < hostMemberVec.size(); ++i) {
             hostMemberVec[i] = i + 0.7;
         }
 
-        FilterBase<TestCell<2> > *filter = new DefaultFilter<TestCell<2>, double, double>();
+        FilterBase<TestCell<2> > *filter = new DefaultFilter<TestCell<2>, float, float>();
 
         filter->copyStreakOut(
             reinterpret_cast<char*>(&hostMemberVec[0]),
@@ -72,7 +72,7 @@ public:
             40);
 
         for (std::size_t i = 0; i < hostBuffer.size(); ++i) {
-            TS_ASSERT_EQUALS(i + 0.7, hostBuffer[i]);
+            TS_ASSERT_EQUALS(float(i + 0.7), hostBuffer[i]);
         }
 
         for (std::size_t i = 0; i < hostBuffer.size(); ++i) {
@@ -88,7 +88,7 @@ public:
             40);
 
         for (std::size_t i = 0; i < hostMemberVec.size(); ++i) {
-            TS_ASSERT_EQUALS(47.11 + i, hostMemberVec[i]);
+            TS_ASSERT_EQUALS(float(47.11 + i), hostMemberVec[i]);
         }
     }
 };
